refactor(tui): colocate Screen::Diff into tui/screens/diff.rs
Fourth screen migration for #287 (Phase 2). Moved into the new
tui/screens/diff.rs: handle_key_diff, diff_guard (the #288 shared guard —
Diff-exclusive, used by both this screen's dispatch and diff_palette_items,
unlike open_config/open_help which are cross-screen entry points),
diff_title, diff_footer, build_diff_vm, render_diff_vm, diff_palette_items.

build_diff_vm is also reused by render_confirm_vm (Screen::Confirm's own
paint, staying in render.rs) as the overwrite-confirm background — its own
doc comment already says so ("also used as Confirm overwrite background").
Left with Diff as its primary owner; the Confirm call site now reaches
across the module boundary (super::screens::diff::build_diff_vm). Likewise
render_diff_pane_vm (render.rs) is shared between Diff's own render and
Confirm's background — stays in render.rs, widened to pub(super).

Pure move — bodies and existing tests unchanged. upload_intent (mod.rs)
stays: shared between List's 'u' and Diff's 'u'. wheel_step's
Diff/Preview/Confirm combined arm (one shared constant across three
screens) is not Diff-exclusive, left untouched.

Part of #287
