+++ src/tui/screens/mod.rs
@@ -5,5 +5,6 @@
 //! keep their exhaustive-match shape; each arm becomes a one-line call into a screen module.
 
 pub(crate) mod config;
+pub(crate) mod diff;
 pub(crate) mod help;
 pub(crate) mod preview;
